+++ CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### CHANGED
+
+- [BUG: 342] NameError: name 'xrange' is not defined on Python 3.8.1
+
 ## [0.11.0] 2020-08-31
 
 ### ADDED
@@ -20,7 +24,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
         - new OptionalFeatures
     - Home
       - accessPointUpdateStates
+	  
 ### CHANGED
+
 - [BUG: 325] Requirements are now using a min version instead of a pinned version. requirements_dev.txt will still use the pinned versions to make sure that the latest version is compatible with the library.
 
 ## [0.10.19] 2020-07-08
@@ -356,6 +362,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [PR: 300]: https://github.com/coreGreenberet/homematicip-rest-api/pull/300
 [PR: 320]: https://github.com/coreGreenberet/homematicip-rest-api/pull/320
 [BUG: 325]: https://github.com/coreGreenberet/homematicip-rest-api/issues/325
+[BUG: 342]: https://github.com/coreGreenberet/homematicip-rest-api/issues/342
 
 [HMIP-FALMOT-C12]: https://github.com/coreGreenberet/homematicip-rest-api/issues/281
 [HMIP-WHS2]: https://github.com/coreGreenberet/homematicip-rest-api/issues/280